requirements.txt
jeepney==0.6.0
loguru==0.5.3
notify-py==0.3.3
pkg-resources==0.0.0
python-crontab==2.5.1
python-dateutil==2.8.1
six==1.16.0
